weight [int--> double] 형변환

+++ src/main/java/com/app/miliwili/src/exercise/dto/PatchExerciseGoalWeight.java
@@ -6,5 +6,5 @@ import lombok.NoArgsConstructor;
 @Getter
 @NoArgsConstructor
 public class PatchExerciseGoalWeight {
-    private Integer goalWeight;
+    private Double goalWeight;
 }
