+++ src/main/java/com/app/miliwili/src/exercise/model/ExerciseInfo.java
@@ -26,11 +26,11 @@ public class ExerciseInfo extends BaseEntity {
 
     @Column(name = "goalWeight")
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer goalWeight;
+    private Double goalWeight;
 
     @Column(name = "firstWeight")
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer firstWeight;
+    private Double firstWeight;
 
     /**
      * User 회원 가입할 시에 필요할 지?
